+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed an issue where websocket updates applied to data that hadn't been initially fetched yet, which resulted in the cache containing only the new partial data instead of all existing data.
+
 ## [0.22.0] - 2026-07-10
 
 ### Changed